Rename everything to BoundingDiameters
