+++ modules/BoundingDiametersSuite/pom.xml
@@ -1,0 +1,61 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <parent>
+        <groupId>org.gephi</groupId>
+        <artifactId>gephi-plugin-parent</artifactId>
+        <version>0.9.2</version>
+    </parent>
+
+    <groupId>nl.liacs</groupId>
+    <artifactId>boundingdiametersssuite</artifactId>
+    <version>0.0.1</version>
+    <packaging>nbm</packaging>
+
+    <name>BoundingDiameters Suite</name>
+    
+    <build>
+        <plugins>
+            <plugin>
+                <groupId>org.codehaus.mojo</groupId>
+                <artifactId>nbm-maven-plugin</artifactId>
+                <configuration>
+                    <licenseName>GPLv3</licenseName>
+                    <author>C. Deve, B. Honig, A. Keizer, S. Lin, D. Nieuwenhuizen, T. Oosterhoorn</author>
+                    <authorEmail></authorEmail>
+                    <authorUrl></authorUrl>
+                    <sourceCodeUrl>http://github.com/alexiooo/gephi-plugins</sourceCodeUrl>
+                    <publicPackages>
+                    </publicPackages>
+                </configuration>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <configuration>
+                    <source>6</source>
+                    <target>6</target>
+                </configuration>
+            </plugin>
+        </plugins>
+    </build>
+    
+    
+    <dependencies>
+        <dependency>
+            <groupId>nl.liacs</groupId>
+            <artifactId>boundingdiameters</artifactId>
+            <version>0.0.1</version>
+        </dependency>
+        <dependency>
+            <groupId>nl.liacs</groupId>
+            <artifactId>boundingdiametersui</artifactId>
+            <version>0.0.1</version>
+        </dependency>
+        <dependency>
+            <groupId>nl.liacs</groupId>
+            <artifactId>boundingdiameterstools</artifactId>
+            <version>0.0.1</version>
+        </dependency>
+    </dependencies>
+</project>
