+++ modules/BoundingDiametersSuite/src/main/nbm/manifest.mf
@@ -1,0 +1,7 @@
+Manifest-Version: 1.0
+OpenIDE-Module-Name: BoundingDiameters Suite
+OpenIDE-Module-Requires: org.openide.windows.WindowManager
+OpenIDE-Module-Short-Description: Plugin implementing the BoundingDiameters algorithm described in the paper "Algorithms for Analyzing and Real-World Graphs"
+OpenIDE-Module-Long-Description: Plugin implementing the BoundingDiameters algorithm described in the paper "Algorithms for Analyzing and Real-World Graphs"
+OpenIDE-Module-Display-Category: Metric
+AutoUpdate-Show-In-Client: false
