+++ modules/BoundingDiametersSuite/src/main/resources/nl/liacs/boundingdiameterssuite/Bundle.properties
