+++ modules/BoundingDiametersUI/src/main/resources/nl/liacs/boundingdiametersui/Bundle.properties
