--- modules/TakesStatisticsTools/pom.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    
-    <parent>
-        <artifactId>gephi-plugin-parent</artifactId>
-        <groupId>org.gephi</groupId>
-        <version>0.9.2</version>
-    </parent>
-    
-    <groupId>nl.liacs</groupId>
-    <artifactId>takesstatisticstools</artifactId>
-    <version>0.0.1</version>
-    <packaging>nbm</packaging>
-
-    <name>Takes Statistics Tools</name>
-
-    <dependencies>
-        <!-- Insert dependencies here -->
-        
-    </dependencies>
-
-    <build>
-        <plugins>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <configuration>
-                    <source>6</source>
-                    <target>6</target>
-                </configuration>
-            </plugin>
-        </plugins>
-    </build>
-</project>
-
-
